% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-narr.R
\name{col2ncvar}
\alias{col2ncvar}
\title{Create \code{ncvar} object from variable name}
\usage{
col2ncvar(variable, dims)
}
\arguments{
\item{variable}{CF variable name}

\item{dims}{List of NetCDF dimension objects (passed to
\code{ncdf4::ncvar_def(..., dim)})}
}
\value{
\code{ncvar} object (from \code{ncvar_def})
}
\description{
Create \code{ncvar} object from variable name
}
