% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coords_prefix}
\alias{coords_prefix}
\title{Format a file name with coordinates}
\usage{
coords_prefix(prefix, extension, coords = umbs_coords(), sep = ".")
}
\arguments{
\item{prefix}{File path and prefix. By default, don't need the trailing \code{.}
because it's set by \code{sep}.}

\item{extension}{File extension. By default, don't need the leading \code{.}
because it's set by \code{sep}.}

\item{coords}{Numeric vector of latitude (name \code{lat}) and longitude (name
\code{lon}). If not named, assume \verb{lat,lon}. Default is \code{\link[=umbs_coords]{umbs_coords()}}.}

\item{sep}{Separator between elements. Default = \code{"."}}
}
\value{
Path with lat-lon included
}
\description{
Format a file name with coordinates
}
