% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-narr.R
\name{download.NARR_site}
\alias{download.NARR_site}
\title{Download NARR time series for a single site}
\usage{
download.NARR_site(
  outfolder,
  start_date,
  end_date,
  lat.in,
  lon.in,
  overwrite = FALSE,
  verbose = FALSE,
  progress = TRUE,
  narr_cache_dir = file.path(outfolder, "narr_cache"),
  overwrite_cache = FALSE,
  parallel = TRUE,
  ncores = parallel::detectCores(),
  silent = TRUE,
  ...
)
}
\arguments{
\item{outfolder}{Target directory for storing output}

\item{start_date}{Start date for met data}

\item{end_date}{End date for met data}

\item{lat.in}{Site latitude coordinate}

\item{lon.in}{Site longitude coordinate}

\item{overwrite}{Overwrite existing files? Default=\code{FALSE}}

\item{verbose}{Turn on verbose output? Default=\code{FALSE}}

\item{progress}{Print a progress bar as data are being downloaded? Default = \code{TRUE}}

\item{narr_cache_dir}{Directory where to cache downloaded data. Default is
\verb{<outfolder>/narr_cache}.}

\item{overwrite_cache}{Delete existing cache? Default = \code{FALSE}}

\item{parallel}{Download in parallel? Default = \code{TRUE}}

\item{ncores}{Number of cores for parallel download. Default is
\code{parallel::detectCores()}}

\item{silent}{Silence error messages?}

\item{...}{Additional arguments (not used)}
}
\description{
Download NARR time series for a single site
}
\examples{

\dontrun{
download.NARR_site(tempdir(), "2001-01-01", "2001-01-12", 43.372, -89.907)
}


}
\author{
Alexey Shiklomanov
}
