% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in.R
\name{ed2in}
\alias{ed2in}
\title{Create (and optionally modify) an ED2IN list based on the default provided in
this package}
\usage{
ed2in(...)
}
\arguments{
\item{...}{Name-value pairs of ED2IN tags to modify. Supports quasiquotation
through \code{\link[rlang:list2]{rlang::list2()}}.}
}
\value{
Named list of tags and values.
}
\description{
Create (and optionally modify) an ED2IN list based on the default provided in
this package
}
