% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortedata2ed.R
\name{fortedata2ed}
\alias{fortedata2ed}
\title{Create ED2 vegetation initial condition files from \code{fortedata} inventory files}
\usage{
fortedata2ed(
  inventory = fortedata::fd_inventory(),
  subplots = fortedata::fd_subplots(),
  output_prefix = NULL
)
}
\arguments{
\item{inventory}{Inventory data. Default is \code{\link[fortedata:fd_inventory]{fortedata::fd_inventory()}}}

\item{subplots}{Subplot information. Default is \code{\link[fortedata:fd_subplots]{fortedata::fd_subplots()}}}

\item{output_prefix}{Output prefix to which to write files. If \code{NULL}
(default), do not write files.}
}
\value{
(Invisibly) List containing cohort (\code{css}) and patch (\code{pss}) \code{tibble}s
}
\description{
Create ED2 vegetation initial condition files from \code{fortedata} inventory files
}
\author{
Alexey Shiklomanov
}
