% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-narr.R
\name{get_NARR_thredds}
\alias{get_NARR_thredds}
\title{Retrieve NARR data using thredds}
\usage{
get_NARR_thredds(
  start_date,
  end_date,
  lat.in,
  lon.in,
  progress = TRUE,
  narr_cache_dir = "narr_cache",
  overwrite_cache = FALSE,
  drop_outside = TRUE,
  parallel = TRUE,
  ncores = 1,
  silent = FALSE
)
}
\arguments{
\item{start_date}{Start date for meteorology}

\item{end_date}{End date for meteorology}

\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}

\item{progress}{Whether or not to show a progress bar (default = \code{TRUE}).
Requires the \code{progress} package to be installed.}

\item{narr_cache_dir}{Directory where to cache downloaded data. Default is
\verb{<outfolder>/narr_cache}.}

\item{overwrite_cache}{Delete existing cache? Default = \code{FALSE}}

\item{drop_outside}{Whether or not to drop dates outside of \code{start_date} to
\code{end_date} range (default = \code{TRUE}).}

\item{parallel}{Download in parallel? Default = \code{TRUE}}

\item{ncores}{Number of cores for parallel download. Default is
\code{parallel::detectCores()}}

\item{silent}{Silence error messages?}
}
\value{
\code{tibble} containing time series of NARR data for the given site
}
\description{
Retrieve NARR data using thredds
}
\examples{

\dontrun{
dat <- get_NARR_thredds("2008-01-01", "2008-01-15", 43.3724, -89.9071)
}

}
\author{
Alexey Shiklomanov
}
