% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ed.R
\name{get_status}
\alias{get_status}
\title{Get the status of an ED2 run}
\usage{
get_status(proc, n = 10)
}
\arguments{
\item{proc}{ED2 process, as returned by \code{\link[=run_ed2]{run_ed2()}}}

\item{n}{Number of output lines to print. Default = 10.}
}
\value{
(Logical) Whether or not the process is still running
}
\description{
Get the status of an ED2 run
}
\author{
Alexey Shiklomanov
}
