% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-narr.R
\name{latlon2narr}
\alias{latlon2narr}
\title{Convert latitude and longitude coordinates to NARR indices}
\usage{
latlon2narr(nc, lat.in, lon.in)
}
\arguments{
\item{nc}{\code{ncdf4} connection object}

\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}
}
\value{
Vector length 2 containing NARR \code{x} and \code{y} indices, which can be
used in \code{ncdf4::ncvar_get} \code{start} argument.
}
\description{
Convert latitude and longitude coordinates to NARR indices
}
\author{
Alexey Shiklomanov
}
