% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{ncget}
\alias{ncget}
\title{Wrapper around \code{ncdf4::ncvar_get} that converts errors and messages to warnings}
\usage{
ncget(nc, varid, ...)
}
\description{
Wrapper around \code{ncdf4::ncvar_get} that converts errors and messages to warnings
}
