% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-xml.R
\name{read_configxml}
\alias{read_configxml}
\title{Read ED2 parameter file (\code{config.xml}) into a \code{data.frame}}
\usage{
read_configxml(filename)
}
\arguments{
\item{filename}{Path and filename of configuration file}
}
\value{
Wide \code{data.frame} containing parameters for each PFT.
}
\description{
Read ED2 parameter file (\code{config.xml}) into a \code{data.frame}
}
