% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in.R
\name{read_ed2in}
\alias{read_ed2in}
\title{Read an ED2IN file to a list}
\usage{
read_ed2in(con, ...)
}
\arguments{
\item{con}{a \link[base]{connection} object or a character string.}

\item{...}{Additional arguments}
}
\value{
Named list of tags and values.
}
\description{
Knows how to handle strings, numbers, and vectors thereof.
}
\author{
Alexey Shiklomanov
}
