% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{read_monthly_dir}
\alias{read_monthly_dir}
\title{Read all monthly output files in a directory}
\usage{
read_monthly_dir(
  outdir,
  save_file = file.path(outdir, "monthly-output.rds"),
  force = FALSE
)
}
\arguments{
\item{outdir}{Directory to scan for files}

\item{save_file}{Name of file for saving output. Defualt is
\code{file.path(outdir, "monthly-output.rds")}}

\item{force}{Logical. If \code{TRUE}, ignore saved file and re-read. If \code{FALSE}
(default), read cached output if it exists, or load output otherwise.}
}
\value{
Nested \code{tibble} with each kind of data returned by \link{read_monthly_file}
}
\description{
Read all monthly output files in a directory
}
