% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{read_monthly_file}
\alias{read_monthly_file}
\title{Read monthly (-E-) file}
\usage{
read_monthly_file(fname, .pb = NULL)
}
\arguments{
\item{fname}{File name}

\item{.pb}{Optional progress bar object}
}
\value{
List of four \code{data.frames}, one each for scalar, cohort, soil, and
PFT variables.
}
\description{
Read monthly (-E-) file
}
