% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species-pfts.R
\name{species_pfts}
\alias{species_pfts}
\title{Conversion betwen FoRTE species and ED2 PFT numbers}
\usage{
species_pfts()
}
\value{
\code{data.frame} with columns \code{Species} (character) and \code{PFT} (double)
}
\description{
Conversion betwen FoRTE species and ED2 PFT numbers
}
\details{
PFT numbers (relevant to FoRTE) are as follows:
\itemize{
\item 5 -- Temperate C3 grass
\item 6 -- Northern pine
\item 8 -- Late-successional conifer
\item 9 -- Early-successional hardwood
\item 10 -- Mid-successional hardwood
\item 11 -- Late-successional hardwood
}
}
\author{
Alexey Shiklomanov
}
