% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in.R
\name{tags2char}
\alias{tags2char}
\title{Format ED2IN tag-value list}
\usage{
tags2char(ed2in)
}
\arguments{
\item{ed2in}{Named list of ED2IN tags (\code{tag = value})}
}
\description{
Converts an \code{ed2in}-like list to an ED2IN-formatted character vector.
}
