% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{umbs_coords}
\alias{umbs_coords}
\alias{umbs_lat}
\alias{umbs_lon}
\title{UMBS coordinates}
\usage{
umbs_coords(digits = 4)

umbs_lat(digits = 4)

umbs_lon(digits = 4)
}
\arguments{
\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}
}
\value{

}
\description{
UMBS coordinates
}
