% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-xml.R
\name{write_configxml}
\alias{write_configxml}
\title{Prepare ED2 parameter config file from a parameter list}
\usage{
write_configxml(param_list, filename = NULL)
}
\arguments{
\item{param_list}{Nested list of parameter values, or a \code{data.frame} with one
row per PFT and parameters in columns (which will be converted to a list
with \code{\link[purrr:transpose]{purrr::transpose()}}).}

\item{filename}{Name of file to write to. Default (\code{NULL}) means don't write
to file.}
}
\value{
In-memory \code{xml2} document representing the file (invisibly)
}
\description{
Prepare ED2 parameter config file from a parameter list
}
\examples{
l <- list(pft = list(num = 9, sla = 22), pft = list(num = 10, sla = 30))
write_configxml(l, tempfile())

# If list has no names, assume that each element is a PFT
# So, below command is identical to above.
l2 <- list(list(num = 9, sla = 22), list(num = 10, sla = 30))
write_configxml(l2, tempfile())

# Can also set non-PFT parameters this way
l3 <- list(
  pft = list(num = 9, sla = 22),
  pft = list(num = 10, sla = 30),
  phenology = list(retained_carbon_fraction = 0.5, theta_crit = 0.3)
)
write_configxml(l3, tempfile())

# data.frames are converted to lists with `purrr::transpose()`
df1 <- data.frame(num = c(9, 10), sla = c(22, 30))
write_configxml(df1, tempfile())
}
\author{
Alexey Shiklomanov
}
