% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in.R
\name{write_ed2in}
\alias{write_ed2in}
\title{Write ED2IN list to file}
\usage{
write_ed2in(ed2in, con, ...)
}
\arguments{
\item{ed2in}{Named list of ED2IN tags (\code{tag = value})}

\item{con}{A \link[base]{connection} object or a character string.}

\item{...}{Additional arguments to \link[base:writeLines]{base::writeLines}}
}
\description{
Write ED2IN list to file
}
