% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2-variables.R
\name{ed2_variable_info}
\alias{ed2_variable_info}
\title{ED2 output variable list}
\usage{
ed2_variable_info(variables = NULL)
}
\arguments{
\item{variables}{Optional character vector of parameters to query}
}
\value{
\code{data.frame} of ED2 variable information
}
\description{
ED2 output variable list
}
\author{
Alexey Shiklomanov
}
