% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ed.R
\name{run_ed2}
\alias{run_ed2}
\title{Run ED2}
\usage{
run_ed2(
  outdir,
  ed2_exe = getOption("ed4forte.ed2_exe"),
  wd = NULL,
  env = NULL,
  stdout = file.path(outdir, "stdout.log"),
  stderr = file.path(outdir, "stderr.log"),
  ...
)
}
\arguments{
\item{outdir}{Output directory}
}
\value{
A \code{\link[processx:process]{processx::process()}} object corresponding to the running ED2
process.
}
\description{
This uses the \code{processx} package to create an ED2 process that runs in the
background. This allows you to do other things while ED2 is running.
}
\author{
Alexey Shiklomanov
}
